55: extract `<Login>` into the newly-created `frontend/src/features/auth/Login.test.tsx` directory
