+++ frontend/src/App.tsx
@@ -17,14 +17,12 @@ import {
   alertsCreate,
 } from "./features/alerts/alertsSlice";
 import { Register } from "./features/auth/Register";
+import { Login } from "./features/auth/Login";
 
 import { PrivateRoute } from "./features/auth/PrivateRoute";
 
 import {
   IState,
-  ActionIssueJWSToken,
-  issueJWSToken,
-  ActionFetchProfile,
   fetchProfile,
   logOut,
   selectHasValidToken,
@@ -191,111 +189,6 @@ export const About = () => {
     <React.Fragment>
       {"<About>"}
       <h1>About VocabTreasury...</h1>
-    </React.Fragment>
-  );
-};
-
-export const Login = () => {
-  console.log(`${new Date().toISOString()} - React is rendering <Login>`);
-
-  const [formData, setFormData] = React.useState({
-    email: "",
-    password: "",
-  });
-
-  const hasValidToken: boolean | null = useSelector(selectHasValidToken);
-  console.log(`    hasValidToken: ${hasValidToken}`);
-
-  const dispatch: ThunkDispatch<
-    IState,
-    unknown,
-    IActionAlertsCreate | ActionIssueJWSToken | ActionFetchProfile
-  > = useDispatch();
-
-  if (hasValidToken === true) {
-    const nextURL: string = "/home";
-    console.log(
-      `    hasValidToken: ${hasValidToken} > redirecting to ${nextURL} ...`
-    );
-    return <Redirect to={nextURL} />;
-  }
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFormData({
-      ...formData,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    const id: string = uuidv4();
-    if (formData.email === "" || formData.password === "") {
-      dispatch(alertsCreate(id, "ALL FORM FIELDS MUST BE FILLED OUT"));
-    } else {
-      try {
-        await dispatch(issueJWSToken(formData.email, formData.password));
-        dispatch(alertsCreate(id, "LOGIN SUCCESSFUL"));
-        await dispatch(fetchProfile());
-      } catch (thunkActionError) {
-        dispatch(alertsCreate(id, thunkActionError));
-      }
-    }
-  };
-
-  return (
-    <React.Fragment>
-      {"<Login>"}
-      <div>
-        <form
-          onSubmit={(e: React.FormEvent<HTMLFormElement>) => handleSubmit(e)}
-        >
-          <fieldset>
-            <legend>[legend-tag: LOG IN]</legend>
-            <div>
-              <label htmlFor="<L>-email">EMAIL</label>
-              <input
-                id="<L>-email"
-                name="email"
-                type="text"
-                value={formData.email}
-                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                  handleChange(e)
-                }
-              />
-            </div>
-            <div>
-              <label htmlFor="<L>-password">PASSWORD</label>
-              <input
-                id="<L>-password"
-                name="password"
-                type="password"
-                value={formData.password}
-                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                  handleChange(e)
-                }
-              />
-            </div>
-          </fieldset>
-          <div>
-            <input
-              id="<L>-submit"
-              name="submit"
-              type="submit"
-              value="LOG INTO MY ACCOUNT"
-            />
-            <small>
-              <Link to="/request_password_reset">FORGOT PASSWORD?</Link>
-            </small>
-          </div>
-        </form>
-      </div>
-      <div>
-        <small>
-          NEED AN ACCOUNT? <Link to="/register">CLICK HERE TO REGISTER</Link>
-        </small>
-      </div>
     </React.Fragment>
   );
 };
